fix: remove unnecessary integer validation

+++ src/resources/team.ts
@@ -1,7 +1,7 @@
 import z from "zod";
 
 export const intraTeamSchema = z.object({
-	id: z.number().int(),
+	id: z.number(),
 	name: z.string(),
 	status: z.string(),
 
@@ -11,9 +11,9 @@ export type intraTeam = z.infer<typeof intraTeamSchema>;
 
 
 export const intraTeamsUserSchema = z.object({
-	id: z.number().int(),
-	team_id: z.number().int(),
-	user_id: z.number().int(),
+	id: z.number(),
+	team_id: z.number(),
+	user_id: z.number(),
 	validated: z.boolean(),
 	leader: z.boolean(),
 });
